refactor(infra): move safe decoder logic
